Add correct methods

+++ WebCbt-Backend/Controllers/EvaluationController.cs
@@ -1,7 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -10,8 +14,10 @@ using WebCbt_Backend.Models;
 
 namespace WebCbt_Backend.Controllers
 {
-    [Route("[controller]")]
     [ApiController]
+    [Authorize]
+    [EnableCors("AllOrigins")]
+    [Route("[controller]")]
     public class EvaluationController : ControllerBase
     {
         private readonly WebCbtDbContext _context;
@@ -21,41 +27,85 @@ namespace WebCbt_Backend.Controllers
             _context = context;
         }
 
-        // GET: api/Evaluations
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<Evaluation>>> GetEvaluations()
+        // POST: /moodtests
+        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [HttpPost("~/moodtests")]
+        public async Task<ActionResult<Evaluation>> PostEvaluation(Evaluation evaluation)
         {
-          if (_context.Evaluations == null)
-          {
-              return NotFound();
-          }
-            return await _context.Evaluations.ToListAsync();
+            if (User.Identity?.IsAuthenticated != true || User.FindFirstValue("userId") != evaluation.UserId.ToString())
+            {
+                return Unauthorized();
+            }
+
+            if (_context.Evaluations == null)
+            {
+                return Problem("Entity set is null.");
+            }
+
+            _context.Evaluations.Add(evaluation);
+
+            await _context.SaveChangesAsync();
+
+            return Ok();
         }
 
-        // GET: api/Evaluations/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Evaluation>> GetEvaluation(int id)
+        // GET: /evaluation/findByUserId?userId=5
+        [HttpGet("findByUserId")]
+        public async Task<ActionResult<IEnumerable<Evaluation>>> FindEvaluationsByUserId(int userId)
         {
-          if (_context.Evaluations == null)
-          {
-              return NotFound();
-          }
-            var evaluation = await _context.Evaluations.FindAsync(id);
+            if (User.Identity?.IsAuthenticated != true)
+            {
+                return Unauthorized();
+            }
+
+            if (_context.Evaluations == null)
+            {
+                return NotFound();
+            }
+
+            return await _context.Evaluations.Where(x => x.UserId == userId).ToListAsync();
+        }
+
+        // GET: /evaluation/5
+        [HttpGet("{evaluationId}")]
+        public async Task<ActionResult<Evaluation>> FindEvaluationById(int evaluationId)
+        {
+            if (User.Identity?.IsAuthenticated != true)
+            {
+                return Unauthorized();
+            }
+
+            if (_context.Evaluations == null)
+            {
+                return NotFound();
+            }
+
+            var evaluation = await _context.Evaluations.FindAsync(evaluationId);
 
             if (evaluation == null)
             {
                 return NotFound();
             }
 
+            if (User.FindFirstValue("userId") != evaluation.UserId.ToString())
+            {
+                return Unauthorized();
+            }
+
             return evaluation;
         }
 
-        // PUT: api/Evaluations/5
+        // PUT: /evaluation/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutEvaluation(int id, Evaluation evaluation)
+        [HttpPut("{evaluationId}")]
+        public async Task<IActionResult> PutEvaluation(int evaluationId, Evaluation evaluation)
         {
-            if (id != evaluation.EvaluationId)
+            if (User.Identity?.IsAuthenticated != true || User.FindFirstValue("userId") != evaluation.UserId.ToString())
+            {
+                return Unauthorized();
+            }
+
+            if (evaluationId != evaluation.EvaluationId)
             {
                 return BadRequest();
             }
@@ -68,7 +118,7 @@ namespace WebCbt_Backend.Controllers
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!EvaluationExists(id))
+                if (!EvaluationExists(evaluationId))
                 {
                     return NotFound();
                 }
@@ -81,44 +131,42 @@ namespace WebCbt_Backend.Controllers
             return NoContent();
         }
 
-        // POST: api/Evaluations
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<Evaluation>> PostEvaluation(Evaluation evaluation)
+        private bool EvaluationExists(int id)
         {
-          if (_context.Evaluations == null)
-          {
-              return Problem("Entity set 'WebCbtDbContext.Evaluations'  is null.");
-          }
-            _context.Evaluations.Add(evaluation);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction("GetEvaluation", new { id = evaluation.EvaluationId }, evaluation);
+            return (_context.Evaluations?.Any(e => e.EvaluationId == id)).GetValueOrDefault();
         }
 
-        // DELETE: api/Evaluations/5
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteEvaluation(int id)
+        // DELETE: /evaluation/5
+        [HttpDelete("{evaluationId}")]
+        public async Task<IActionResult> DeleteEvaluation(int evaluationId)
         {
+            if (User.Identity?.IsAuthenticated != true)
+            {
+                return Unauthorized();
+            }
+
             if (_context.Evaluations == null)
             {
                 return NotFound();
             }
-            var evaluation = await _context.Evaluations.FindAsync(id);
+
+            var evaluation = await _context.Evaluations.FindAsync(evaluationId);
+
             if (evaluation == null)
             {
                 return NotFound();
             }
 
+            if (User.FindFirstValue("userId") != evaluation.UserId.ToString())
+            {
+                return Unauthorized();
+            }
+
             _context.Evaluations.Remove(evaluation);
+
             await _context.SaveChangesAsync();
 
             return NoContent();
-        }
-
-        private bool EvaluationExists(int id)
-        {
-            return (_context.Evaluations?.Any(e => e.EvaluationId == id)).GetValueOrDefault();
         }
     }
 }
